+++ packages/apollo-shared/src/Changes/MergeExonsChange.ts
@@ -1,4 +1,3 @@
-/* eslint-disable unicorn/prefer-structured-clone */
 /* eslint-disable @typescript-eslint/require-await */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 
@@ -10,10 +9,7 @@ import {
   SerializedFeatureChange,
   ServerDataStore,
 } from '@apollo-annotation/common'
-import {
-  AnnotationFeature,
-  AnnotationFeatureSnapshot,
-} from '@apollo-annotation/mst'
+import { AnnotationFeatureSnapshot } from '@apollo-annotation/mst'
 import { findAndDeleteChildFeature } from './DeleteFeatureChange'
 import { SplitExonChange } from './SplitExonChange'
 
@@ -52,7 +48,9 @@ export class MergeExonsChange extends FeatureChange {
   toJSON(): SerializedMergeExonsChange {
     const { assembly, changedIds, changes, typeName } = this
     if (changes.length === 1) {
-      const [{ firstExon, secondExon, parentFeatureId, mergedExon }] = changes
+      const [
+        { firstExon, secondExon, parentFeatureId, AnnotationFeatureSnapshot },
+      ] = changes
 
       return {
         typeName,
@@ -61,7 +59,6 @@ export class MergeExonsChange extends FeatureChange {
         firstExon,
         secondExon,
         parentFeatureId,
-        mergedExon,
       }
     }
     return { typeName, changedIds, assembly, changes }
@@ -71,7 +68,7 @@ export class MergeExonsChange extends FeatureChange {
     const { featureModel, session } = backend
     const { changes, logger } = this
     for (const change of changes) {
-      const { firstExon, secondExon, mergedExon } = change
+      const { firstExon, secondExon } = change
       const topLevelFeature = await featureModel
         .findOne({ allIds: firstExon._id })
         .session(session)
@@ -81,19 +78,15 @@ export class MergeExonsChange extends FeatureChange {
         logger.error(errMsg)
         throw new Error(errMsg)
       }
-
-      const updatedExon = this.getFeatureFromId(topLevelFeature, firstExon._id)
-      if (!updatedExon) {
+      const mergedExon = this.getFeatureFromId(topLevelFeature, firstExon._id)
+      if (!mergedExon) {
         const errMsg = 'ERROR when searching feature by featureId'
         logger.error(errMsg)
         throw new Error(errMsg)
       }
-      updatedExon.min = mergedExon.min
-      updatedExon.max = mergedExon.max
-      updatedExon.attributes = JSON.parse(
-        JSON.stringify(mergedExon.attributes),
-      ) as Record<string, string[]>
-
+      mergedExon.min = Math.min(firstExon.min, secondExon.min)
+      mergedExon.max = Math.max(firstExon.max, secondExon.max)
+      mergedExon.attributes = this.mergeAttributes(firstExon, secondExon)
       const deletedIds = findAndDeleteChildFeature(
         topLevelFeature,
         secondExon._id,
@@ -118,21 +111,19 @@ export class MergeExonsChange extends FeatureChange {
     }
 
     for (const [idx, changedId] of this.changedIds.entries()) {
-      const { firstExon, secondExon, mergedExon } = this.changes[idx]
-      const updatedExon = dataStore.getFeature(firstExon._id)
-      if (!updatedExon) {
+      const { firstExon, secondExon } = this.changes[idx]
+      const mergedExon = dataStore.getFeature(firstExon._id)
+      if (!mergedExon) {
         throw new Error(`Could not find feature with identifier "${changedId}"`)
       }
-      updatedExon.setMin(mergedExon.min)
-      updatedExon.setMax(mergedExon.max)
-      const mergedAttrs = JSON.parse(
-        JSON.stringify(mergedExon.attributes),
-      ) as Record<string, string[]>
-      updatedExon.setAttributes(new Map())
+      mergedExon.setMin(Math.min(firstExon.min, secondExon.min))
+      mergedExon.setMax(Math.max(firstExon.max, secondExon.max))
+      const mergedAttrs = this.mergeAttributes(firstExon, secondExon)
+      mergedExon.setAttributes(new Map())
       for (const key of Object.keys(mergedAttrs)) {
-        updatedExon.setAttribute(key, mergedAttrs[key])
+        mergedExon.setAttribute(key, mergedAttrs[key])
       }
-      updatedExon.parent?.deleteChild(secondExon._id)
+      mergedExon.parent?.deleteChild(secondExon._id)
     }
   }
 
@@ -176,38 +167,39 @@ export class MergeExonsChange extends FeatureChange {
       upstreamCut,
       downstreamCut,
     }
-    // console.log('inverseChanges:' + JSON.stringify(inverseChange, null, 2))
+    console.log('inverseChanges:' + JSON.stringify(inverseChange, null, 2))
     return inverseChange
   }
 
-  // mergeAttributes(
-  //   firstExon: AnnotationFeatureSnapshot,
-  //   secondExon: AnnotationFeatureSnapshot,
-  // ): Record<string, string[]> {
-  //   let mergedAttrs: Record<string, string[]> = {}
-  //   if (firstExon.attributes) {
-  //     // eslint-disable-next-line unicorn/prefer-structured-clone
-  //     mergedAttrs = JSON.parse(JSON.stringify(firstExon.attributes))
-  //   }
+  mergeAttributes(
+    firstExon: AnnotationFeatureSnapshot,
+    secondExon: AnnotationFeatureSnapshot,
+  ): Record<string, string[]> {
+    let mergedAttrs: Record<string, string[]> = {}
+    if (firstExon.attributes) {
+      // eslint-disable-next-line unicorn/prefer-structured-clone
+      mergedAttrs = JSON.parse(JSON.stringify(firstExon.attributes))
+    }
 
-  //   if (secondExon.attributes) {
-  //     const attrs: Record<string, string[]> = JSON.parse(
-  //       JSON.stringify(secondExon.attributes),
-  //     )
-  //     for (const key of Object.keys(attrs)) {
-  //       if (key === '_id' || key === 'gff_id') {
-  //         continue
-  //       }
-  //       if (!Object.keys(mergedAttrs).includes(key)) {
-  //         mergedAttrs[key] = []
-  //       }
-  //       attrs[key].map((x) => {
-  //         if (!mergedAttrs[key].includes(x)) {
-  //           mergedAttrs[key].push(x)
-  //         }
-  //       })
-  //     }
-  //   }
-  //   return mergedAttrs
-  // }
+    if (secondExon.attributes) {
+      // eslint-disable-next-line unicorn/prefer-structured-clone
+      const attrs: Record<string, string[]> = JSON.parse(
+        JSON.stringify(secondExon.attributes),
+      )
+      for (const key of Object.keys(attrs)) {
+        if (key === '_id' || key === 'gff_id') {
+          continue
+        }
+        if (!Object.keys(mergedAttrs).includes(key)) {
+          mergedAttrs[key] = []
+        }
+        attrs[key].map((x) => {
+          if (!mergedAttrs[key].includes(x)) {
+            mergedAttrs[key].push(x)
+          }
+        })
+      }
+    }
+    return mergedAttrs
+  }
 }
